A python module to import Mitchell Manager shop data
pymitchell is a module to import data from mitchell manager shop
software into easier to use objects

pymitchell uses a sqlite database which is made from the original
SMCORE32.mdb database that mitchell uses. The table names in sqlite
database have a preceding '_' before them, this was done because sqlite
didn't like  some of the original names. The key names have a trailing
'_' added to them for the same reason.